Renamed loss to costs
