+++ pkg/costs/kind.go
@@ -1,4 +1,4 @@
-package loss
+package costs
 
 type Kind int
