+++ pkg/costs/mse.go
@@ -1,4 +1,4 @@
-package loss
+package costs
 
 import (
 	"fmt"
@@ -9,7 +9,7 @@ type MSECostFunction struct{}
 
 func (costFunction *MSECostFunction) Evaluate(actual mat.Vector) int {
 	if actual == nil {
-		panic("loss: mse cost function got nil vector")
+		panic("costs: mse cost function got nil vector")
 	}
 	index, max := 0, actual.AtVec(0)
 	for i := 1; i < actual.Len(); i++ {
@@ -23,11 +23,11 @@ func (costFunction *MSECostFunction) Evaluate(actual mat.Vector) int {
 
 func (costFunction *MSECostFunction) Differentiate(actual mat.Vector, label int) mat.Vector {
 	if actual == nil {
-		panic("loss: mse cost function got nil vector")
+		panic("costs: mse cost function got nil vector")
 	}
 	length := actual.Len()
 	if label < 0 || label >= length {
-		panic(fmt.Sprintf("loss: mse cost function got invalid label, %d", label))
+		panic(fmt.Sprintf("costs: mse cost function got invalid label, %d", label))
 	}
 	costs := mat.NewVecDense(length, nil)
 	costs.SetVec(label, 1)
